Add SPACESHIP_DIR_TRUNC_PREFIX, a prefix before cwd when truncated
When the cwd is truncated by SPACESHIP_DIR_TRUNC instead of showing:

    a/b/c

It will show:

    .../a/b/c

The prefix is configurable of course and can be disabled by setting
it to an empty string.

I also modified SPACESHIP_DIR_TRUNC_REPO to use
SPACESHIP_DIR_TRUNC_PREFIX and made it not truncate on a repository that
is below the root directory. In addition, I optimized its code to use
Zsh constructs instead of command substitutions.